security

+++ src/utils/supabase/tenant/actions.ts
@@ -11,7 +11,6 @@ export async function insertNewTenant(formData: FormData) {
     process.env.NEXT_PUBLIC_SUPABASE_URL || '',
     process.env.SUPABASE_SERVICE_ROLE_KEY || ''
   );
-  //const supabase = createClient()
 
   const email = String(formData.get('email'))
   const full_name = String(formData.get('fullName'))
@@ -24,16 +23,6 @@ export async function insertNewTenant(formData: FormData) {
   const emailRegex = /^[a-zA-Z0-9._-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/;
   if (!emailRegex.test(email.toLowerCase())) throw new Error('Invalid email')
   
-  // const { data, error } = await supabaseAdmin.from('users')
-  //   .select('*')
-  //   .eq('email', email)
-  //   .single()
-  // const user = await getUser()
-  // console.log("USERSEURSEURUUSER************************************", user)
-  // if (!user) throw new Error('User not found')
-  // if (error || !data) throw new Error(error?.message || 'User not found')
-  
-  // console.log(data)
   const { error: insertError } = await supabaseAdmin.from('tenants').upsert({
     email: email.toLowerCase(),
     property_id: propertyId,
@@ -56,11 +45,11 @@ export async function deleteInvite(formData: FormData) {
   const token = String(formData.get('token'))
   if (!token) return
 
-  const { data } = await supabaseAdmin.from('property_invites')
+  const { data, error: inviteError } = await supabaseAdmin.from('property_invites')
     .select('*')
     .eq('token', token)
   
-  if (!data || data.length < 1) throw new Error('Invite not found')  
+  if (inviteError || !data || data.length < 1) throw new Error('Invite not found')  
 
   const { error } = await supabaseAdmin.from('property_invites')
     .delete()
